LayoutTests/fast/css/counters/content-021-expected.txt

This is the WebKit version of CSS Test: Content using a 'counters()' function with a string value.

PASS layoutTestController.counterValueForElementById('div1') is '0'
PASS layoutTestController.counterValueForElementById('div2') is '0.0'

TEST COMPLETE
